let listbox cmd to show the alias if having
